+++ layers/GatoGraphQLForWP/phpunit-packages/webserver-requests/tests/AbstractFixtureThirdPartyPluginDependencyWordPressAuthenticatedUserWebserverRequestTestCase.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace PHPUnitForGatoGraphQL\WebserverRequests;
 
 use GraphQLByPoP\GraphQLServer\Unit\FixtureTestCaseTrait;
+use PHPUnitForGatoGraphQL\GatoGraphQL\Integration\FixtureWebserverRequestTestCaseTrait;
 
 use function file_get_contents;
 
@@ -14,6 +15,7 @@ use function file_get_contents;
 abstract class AbstractFixtureThirdPartyPluginDependencyWordPressAuthenticatedUserWebserverRequestTestCase extends AbstractThirdPartyPluginDependencyWordPressAuthenticatedUserWebserverRequestTestCase
 {
     use FixtureTestCaseTrait;
+    use FixtureWebserverRequestTestCaseTrait;
 
     /**
      * Since PHPUnit v10, this is not possible anymore!
@@ -53,11 +55,8 @@ abstract class AbstractFixtureThirdPartyPluginDependencyWordPressAuthenticatedUs
             if (!\file_exists($pluginDisabledGraphQLResponseFile)) {
                 static::throwFileNotExistsException($pluginDisabledGraphQLResponseFile);
             }
-            $pluginOnlyOneEnabledGraphQLResponse = null;
             $pluginOnlyOneEnabledGraphQLResponseFile = $filePath . \DIRECTORY_SEPARATOR . $fileName . ':only-one-enabled.json';
-            if (\file_exists($pluginOnlyOneEnabledGraphQLResponseFile)) {
-                $pluginOnlyOneEnabledGraphQLResponse = file_get_contents($pluginOnlyOneEnabledGraphQLResponseFile);
-            }
+            $pluginGraphQLVariablesFile = $filePath . \DIRECTORY_SEPARATOR . $fileName . '.var.json';
 
             // The plugin name is created by the folder (plugin vendor) + fileName (plugin name)
             $pluginVendor = substr($filePath, strlen($fixtureFolder . '/'));
@@ -67,8 +66,50 @@ abstract class AbstractFixtureThirdPartyPluginDependencyWordPressAuthenticatedUs
                 'response-enabled' => file_get_contents($pluginEnabledGraphQLResponseFile),
                 'response-disabled' => file_get_contents($pluginDisabledGraphQLResponseFile),
             ];
-            if ($pluginOnlyOneEnabledGraphQLResponse !== null) {
-                $pluginEntries[$pluginName]['response-only-one-enabled'] = $pluginOnlyOneEnabledGraphQLResponse;
+            if (\file_exists($pluginOnlyOneEnabledGraphQLResponseFile)) {
+                $pluginEntries[$pluginName]['response-only-one-enabled'] = file_get_contents($pluginOnlyOneEnabledGraphQLResponseFile);
+            }
+            if (\file_exists($pluginGraphQLVariablesFile)) {
+                $pluginEntries[$pluginName]['variables'] = static::getGraphQLVariables($pluginGraphQLVariablesFile);
+
+                /**
+                 * Check if there are additional response entries (with different vars)
+                 * for the same query. For that, search for all items of type :{number}.var.json,
+                 * and add those entries with the same query.
+                 */
+                $additionalPluginGraphQLVariablesFiles = static::findFilesInDirectory(
+                    $filePath,
+                    [$fileName . ':*.var.json'],
+                    ['*.disabled.var.json']
+                );
+                foreach ($additionalPluginGraphQLVariablesFiles as $additionalPluginGraphQLVariablesFile) {
+                    $additionalFileName = $additionalPluginGraphQLVariablesFile->getFilenameWithoutExtension();
+                    $additionalFileNumberWithSuffix = substr($additionalFileName, strpos($additionalFileName, ':') + 1);
+                    /** @var int */
+                    $dotVarPos = strpos($additionalFileNumberWithSuffix, '.var');
+                    $additionalFileNumber = substr($additionalFileNumberWithSuffix, 0, $dotVarPos);
+                    if (!is_numeric($additionalFileNumber)) {
+                        continue;
+                    }
+                    $additionalPluginName = $pluginName . ':' . $additionalFileNumber;
+                    $pluginEntries[$additionalPluginName] = [
+                        'query' => $query,
+                        'variables' => static::getGraphQLVariables($additionalPluginGraphQLVariablesFile->getRealPath()),
+                    ];
+                    // For the variations, make the "enable" and "disable" responses optional
+                    $additionalPluginEnabledGraphQLResponseFile = $filePath . \DIRECTORY_SEPARATOR . $fileName . ':' . $additionalFileNumber . ':enabled.json';
+                    if (\file_exists($additionalPluginEnabledGraphQLResponseFile)) {
+                        $pluginEntries[$additionalPluginName]['response-enabled'] = file_get_contents($additionalPluginEnabledGraphQLResponseFile);
+                    }
+                    $additionalPluginDisabledGraphQLResponseFile = $filePath . \DIRECTORY_SEPARATOR . $fileName . ':' . $additionalFileNumber . ':disabled.json';
+                    if (\file_exists($additionalPluginDisabledGraphQLResponseFile)) {
+                        $pluginEntries[$additionalPluginName]['response-disabled'] = file_get_contents($additionalPluginDisabledGraphQLResponseFile);
+                    }
+                    $additionalPluginOnlyOneEnabledGraphQLResponseFile = $filePath . \DIRECTORY_SEPARATOR . $fileName . ':only-one-enabled.json';
+                    if (\file_exists($additionalPluginOnlyOneEnabledGraphQLResponseFile)) {
+                        $pluginEntries[$additionalPluginName]['response-only-one-enabled'] = file_get_contents($additionalPluginOnlyOneEnabledGraphQLResponseFile);
+                    }
+                }
             }
         }
         return static::customizePluginNameEntries($pluginEntries);