+++ layers/GatoGraphQLForWP/plugins/testing-schema/gatographql-testing-schema.php
@@ -3,7 +3,7 @@
 Plugin Name: Gato GraphQL - Testing Schema
 Plugin URI: https://github.com/GatoGraphQL/GatoGraphQL
 Description: Addition of elements to the GraphQL schema to test the Gato GraphQL plugin
-Version: 3.1.0-dev
+Version: 4.0.0-dev
 Requires at least: 6.0
 Requires PHP: 8.1
 Author: Gato GraphQL
@@ -52,7 +52,7 @@ add_action(
          *
          * @gatographql-readonly-code
          */
-        $extensionVersion = '3.1.0-dev';
+        $extensionVersion = '4.0.0-dev';
         $extensionName = \__('Gato GraphQL - Testing Schema', 'gatographql-testing-schema');
         /**
          * Important: Do not modify the formatting of this PHP code!
@@ -63,7 +63,7 @@ add_action(
          *
          * @gatographql-readonly-code
          */
-        $mainPluginVersionConstraint = '^3.1';
+        $mainPluginVersionConstraint = '^4.0';
         
         /**
          * Validate Gato GraphQL is active