+++ src/Monorepo/MonorepoMetadata.php
@@ -16,7 +16,7 @@ final class MonorepoMetadata
      *
      * @gatographql-readonly-code
      */
-    final public const VERSION = '3.1.0-dev';
+    final public const VERSION = '4.0.0-dev';
     /**
      * This const will reflect the latest published tag in GitHub.
      *