+++ layers/GatoGraphQLForWP/plugins/gatographql/CHANGELOG.md
@@ -12,7 +12,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Added documentation for new PRO module **Polylang Mutations** (#2733)
 - Added documentation for new PRO field `_arrayIntersect` (#2735)
 - Added predefined persisted query:
-  - [PRO] Create missing translation posts for Polylang
+  - [PRO] Create missing translation posts for Polylang (#2740)
 
 ### Fixed